IdCode

+++ models/userModel.js
@@ -1,18 +1,8 @@
 // userModel.js
 
 const dbConnection = require('../config/dbConfig');
-const Filter = require('bad-words');
-const filter = new Filter();
-
-// Function to generate hex code ID
-const generateHexCode = (length) => {
-  const characters = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ';
-  let result = '';
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * characters.length));
-  }
-  return result;
-};
+const { generateIdCode, isIdCode } = require('../utils/idCodeGenerator');
+const bcrypt = require('bcrypt');
 
 // Function to create a user
 const createUser = async (userData) => {
@@ -24,34 +14,47 @@ const createUser = async (userData) => {
     if (!userData.email) {
         throw new Error('Email is a required field.');
     }
+
+    if (!userData.password) {
+        throw new Error('Password is a required field.');
+    }
   
     // Check if email format is valid
     const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
     if (!emailRegex.test(userData.email)) {
       throw new Error('Invalid email format.');
     }
+    // Check if email is already registered
+    const existingEmail = await dbConnection.promise().query(
+        'SELECT * FROM users WHERE email = ?',
+        [userData.email]
+    );
+    if (existingEmail[0].length > 0){
+        throw new Error('An account with this email already exists.');
+    }
 
     
 
     // Check if username is already taken
     const existingUsername = await dbConnection.promise().query(
-        'SELECT count(*) FROM users WHERE username = ?',
+        'SELECT * FROM users WHERE username = ?',
         [userData.username]
     );
-    if (existingUsername[0] != 0) {
+    if (existingUsername[0].length > 0){
         throw new Error('Username is already taken.');
     }
-    if (filter.isProfane(userData.username)) {
-        throw new Error('Username contains prohibited words.');
+    // Check if username contains only lowercase letters and numbers
+    const usernameRegex = /^[a-z0-9]+$/;
+    if (!usernameRegex.test(userData.username)) {
+        throw new Error('Username can only contain lowercase letters and numbers.');
     }
-
 
 
     // Check if user ID is already taken (assuming user ID is unique)
     let userId;
     let existingUserId;
     do {
-        userId = generateHexCode(8); // Generate 8 character hex code for user ID
+        userId = generateIdCode();
         existingUserId = await dbConnection.promise().query(
             'SELECT * FROM users WHERE user_id = ?',
             [userId]
@@ -67,7 +70,6 @@ const createUser = async (userData) => {
     const hashedPassword = await bcrypt.hash(userData.password, 10); // Use salt rounds of 10
   
     try {
-        const userId = generateHexCode(8); // Generate 8 character hex code for user ID
         const [rows, fields] = await dbConnection.promise().execute(
             'INSERT INTO users (user_id, username, full_name, email, password, register_date, update_date) VALUES (?, ?, ?, ?, ?, NOW(), NOW())',
             [userId, userData.username, userData.fullName, userData.email, hashedPassword]
@@ -80,6 +82,10 @@ const createUser = async (userData) => {
   
 
 const deleteUserById = async (userId) => {
+    if (!isIdCode(userId)) {
+        throw new Error('Invalid user ID.');
+    }
+
     try {
         const [rows, fields] = await dbConnection.promise().execute(
         'DELETE FROM users WHERE user_id = ?',
@@ -100,5 +106,41 @@ const editUserById = async (userId, userData) => {
         throw new Error(err.message);
     }
 };
+
+const getAll = async (userId, userData) => {
+    try {
+        const [userRows, fields] = await dbConnection.promise().query(
+            'SELECT user_id id, username, full_name, register_date FROM users ORDER BY register_date DESC'
+        );
+
+        const [countRows, _] = await dbConnection.promise().query(
+            'SELECT COUNT(*) AS user_count FROM users'
+        );
+
+        const totalCount = countRows[0].user_count;
+
+        return {
+            users: userRows,
+            totalCount: totalCount
+        };
+    } catch (err) {
+        throw new Error(err.message);
+    }
+};
+
+const getUserById = async (userId) => {
+    try {
+        const [rows, fields] = await dbConnection.promise().query(
+        'SELECT user_id id, username, full_name, register_date FROM users WHERE user_id = ?',
+        [userId]
+        );
+        if (rows.length === 0) {
+        throw new Error('User not found.');
+        }
+        return rows[0];
+    } catch (err) {
+        throw new Error(err.message);
+    }
+};
   
-  module.exports = { createUser, deleteUserById, editUserById };
+  module.exports = { createUser, deleteUserById, editUserById, getAll, getUserById };
